Rename contracts -> contract, move contract to workdir, add an example of contracts usage
