+++ src/modules/bootstrap/README.md
@@ -1,3 +1,3 @@
 # @prescott/bootstrap
-Implementation of root config loader for Prescott.\
-Bootstraps application according to provided configuration and checks system consistency.
+Implementation of root config loader for Prescott.
+Bootstraps application according to provided configuration.
