+++ src/modules/bootstrap/schema/root-config.schema.ts
@@ -1,5 +1,5 @@
 import * as Joi from 'joi';
-import { contractConfigSchema } from '@modules/contracts/schema/contract-config.schema';
+import { contractConfigSchema } from '@modules/contract/schema/contract-config.schema';
 import { RootConfigFile } from '@modules/bootstrap/model/root-config';
 
 export const rootConfigFileSchema = Joi.object<RootConfigFile>({
