+++ src/modules/contract/model/contract.d.ts
