+++ src/modules/contract/model/log-entry.d.ts
