+++ src/modules/contract/model/metric-entry.ts
