+++ src/modules/contract/schema/log-provider.schema.ts
@@ -1,5 +1,5 @@
 import * as Joi from 'joi';
-import { LogProviderContract } from '@modules/contracts/model/log-provider.contract';
+import { LogProviderContract } from '@modules/contract/model/log-provider.contract';
 
 export const logProviderSchema = Joi.object<LogProviderContract>({
   init: Joi.function().minArity(1).required(),
