+++ src/workdir/README.md
@@ -1,0 +1,4 @@
+# Workdir
+This directory exists only for development purposes. In a real application, it will be located
+in some other place like `/var/lib/prescott`, not within a code.
+**NOTE**: all implementations in `./contract` should be used by default.
