+++ src/workdir/config.json
@@ -2,15 +2,15 @@
   "contract": {
     "env": {
       "type": "file",
-      "key": "../modules/contracts/impl/env-provider.ts"
+      "key": "env-provider.ts"
     },
     "log": {
       "type": "file",
-      "key": "../modules/contracts/impl/log-provider.ts"
+      "key": "log-provider.ts"
     },
     "metric": {
       "type": "file",
-      "key": "../modules/contracts/impl/metric-provider.ts"
+      "key": "metric-provider.ts"
     }
   }
 }
